+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/stream/RecorderStateStreamHandler.kt
@@ -2,12 +2,14 @@ package com.llfbandit.record.record.stream
 
 import android.os.Handler
 import android.os.Looper
+import com.llfbandit.record.record.RecordState
 import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.EventChannel.EventSink
 
 class RecorderStateStreamHandler : EventChannel.StreamHandler {
     // Event producer
     private var eventSink: EventSink? = null
+    private var state: RecordState = RecordState.STOP
 
     private val uiThreadHandler = Handler(Looper.getMainLooper())
 
@@ -19,9 +21,13 @@ class RecorderStateStreamHandler : EventChannel.StreamHandler {
         eventSink = null
     }
 
-    fun sendStateEvent(state: Int) {
-        uiThreadHandler.post {
-            eventSink?.success(state)
+    fun sendStateEvent(state: RecordState) {
+        if (this.state != state) {
+            this.state = state
+
+            uiThreadHandler.post {
+                eventSink?.success(state.id)
+            }
         }
     }