+++ record_android/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.3.2
+* fix: Pause state accuracy.
+
 ## 1.3.1
 * fix: Stop does not complete when called immediately after startStream.
 * fix: No static method `addLast`.